Add C-k keybinding to windowlist to kill windows without confirmation
'K' in windowlist will ask you if you want to kill the currently
selected window and if you say yes it will kill it and scroll to the top
buffer. This makes killing many windows annoying because you are
constantly confirming and scrolling back down the list.

This change adds the keybinding 'C-k' that will kill the current window
without asking you for confirmation and will attempt to keep the
selection in the same area of the windowlist as the killed window was.

The last part, handling what is selected in the windowlist after the
window is killed, is not correctly handled for the last or current
buffer. For the last buffer, the first item in the list is selected. For
the current window, the windowlist exits. Patches to fix these are
welcome.